Add a template for a Lambda@Edge function
This CloudFormation template provisions a Lambda function that should be associated with a CloudFront cache behavior's origin requests. The function effectively rewrites any origin request that ends with a `/` to a request that ends with `/index.html`. The index document to use is a parameter of the template and thus can be configured when the function is provisioned.