Add a template for a static website
This CloudFormation template provisions an S3 bucket to host the static assets, a bucket policy that requires a CloudFront origin access identity to access the website, a CloudFront distribution with an S3 origin and a default cache behavior, and a Route 53 zone with A record(s). Provisioning the CDN and DNS zone requires a valid ACM certificate. A traditional www host name can be configured so that the website can be hosted on example.com and www.example.com (where the first redirects to the second or vice versa). All requests for the website are redirected to HTTPS (using a single redirect) and requests for paths ending in `/` can optionally be rewritten to requests for an index document (using the `IndexDocument` Lambda@Edge function found in this repository). The content of error responses can optionally be replaced by the content of an error document.